back/requirements.txt
Django~=4.2.7
django-cors-headers==4.6.0
psycopg2-binary==2.9.11
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
django-formset==1.5.5
drf-yasg==1.21.8
PyJWT==2.9.0
requests
python-dotenv
django-filter~=24.3
djangorestframework-simplejwt
drf-spectacular
django-filter~=24.3